+++ ui/packages/app/web/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.895](https://github.com/parca-dev/parca/compare/@parca/web@0.16.894...@parca/web@0.16.895) (2025-07-08)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.894](https://github.com/parca-dev/parca/compare/@parca/web@0.16.893...@parca/web@0.16.894) (2025-07-08)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.893](https://github.com/parca-dev/parca/compare/@parca/web@0.16.892...@parca/web@0.16.893) (2025-07-08)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.892](https://github.com/parca-dev/parca/compare/@parca/web@0.16.891...@parca/web@0.16.892) (2025-07-03)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.891](https://github.com/parca-dev/parca/compare/@parca/web@0.16.890...@parca/web@0.16.891) (2025-07-02)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.890](https://github.com/parca-dev/parca/compare/@parca/web@0.16.889...@parca/web@0.16.890) (2025-06-28)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.889](https://github.com/parca-dev/parca/compare/@parca/web@0.16.888...@parca/web@0.16.889) (2025-06-28)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.888](https://github.com/parca-dev/parca/compare/@parca/web@0.16.887...@parca/web@0.16.888) (2025-06-26)
+
+**Note:** Version bump only for package @parca/web
+
 ## [0.16.887](https://github.com/parca-dev/parca/compare/@parca/web@0.16.886...@parca/web@0.16.887) (2025-06-25)
 
 **Note:** Version bump only for package @parca/web