+++ ui/packages/shared/hooks/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.92 (2025-07-08)
+
+**Note:** Version bump only for package @parca/hooks
+
 ## 0.0.91 (2025-06-18)
 
 **Note:** Version bump only for package @parca/hooks